+++ src/lib.rs
@@ -4,7 +4,7 @@
 //! # snake_bevy_wasm
 //!
 //! **Simple snake game with Bevy, Rust and Wasm**  
-//! ***version: 1.3.41 date: 2025-08-08 author: [bestia.dev](https://bestia.dev) repository: [GitHub](https://github.com/bestia-dev/snake_bevy_wasm)***
+//! ***version: 1.3.44 date: 2025-08-08 author: [bestia.dev](https://bestia.dev) repository: [GitHub](https://github.com/bestia-dev/snake_bevy_wasm)***
 //!
 //!  ![maintained](https://img.shields.io/badge/maintained-green)
 //!  ![work-in-progress](https://img.shields.io/badge/work_in_progress-yellow)
@@ -14,9 +14,9 @@
 //!  [![License](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/bestia-dev/snake_bevy_wasm/blob/master/LICENSE)
 //!  ![snake_bevy_wasm](https://bestia.dev/webpage_hit_counter/get_svg_image/1481465721.svg)
 //!
-//! [![Lines in Rust code](https://img.shields.io/badge/Lines_in_Rust-1333-green.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
-//! [![Lines in Doc comments](https://img.shields.io/badge/Lines_in_Doc_comments-84-blue.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
-//! [![Lines in Comments](https://img.shields.io/badge/Lines_in_comments-140-purple.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
+//! [![Lines in Rust code](https://img.shields.io/badge/Lines_in_Rust-1332-green.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
+//! [![Lines in Doc comments](https://img.shields.io/badge/Lines_in_Doc_comments-82-blue.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
+//! [![Lines in Comments](https://img.shields.io/badge/Lines_in_comments-138-purple.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
 //! [![Lines in examples](https://img.shields.io/badge/Lines_in_examples-0-yellow.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
 //! [![Lines in tests](https://img.shields.io/badge/Lines_in_tests-0-orange.svg)](https://github.com/bestia-dev/snake_bevy_wasm/)
 //!
@@ -113,9 +113,6 @@ struct CustomMaterial {
 /// The Material trait is very configurable, but comes with sensible defaults for all methods.
 /// You only need to implement functions for features that need non-default behavior. See the Material api docs for details!
 impl Material2d for CustomMaterial {
-    fn vertex_shader() -> ShaderRef {
-        "custom_material.wgsl".into()
-    }
     fn fragment_shader() -> ShaderRef {
         "custom_material.wgsl".into()
     }